refactor: move window types to own file
