+++ src/components/Notepad/Notepad.tsx
@@ -15,8 +15,8 @@ import BaseWindow from "../BaseWindow/BaseWindow";
 import {
   useWindowState,
   WindowConfig,
-  WindowType,
 } from "../WindowManager/window-management-slice";
+import { WindowType } from "../WindowManager/window-type";
 import { useDispatch } from "react-redux";
 
 // CSS
